Add code review fixes
