+++ HappyTravel.Hiroshima.Common/Models/Accommodations/Rooms/CancellationPolicies/Policy.cs
