+++ HappyTravel.Hiroshima.Common/Models/Accommodations/Rooms/CancellationPolicies/PolicyPenaltyTypes.cs
@@ -1,6 +1,6 @@
 ﻿namespace HappyTravel.Hiroshima.Common.Models.Accommodations.Rooms.CancellationPolicies
 {
-    public enum CancellationPenaltyTypes
+    public enum PolicyPenaltyTypes
     {
         Percent,
         Nights
